mini_projects/benke.tibor/backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ASGI server
uvicorn==0.27.0
# uvloop==0.19.0  # Windows incompatible - use standard asyncio instead

# LangChain + LangGraph
langchain==0.1.11
langchain-openai==0.0.8
langgraph==0.0.20
langchain-community>=0.0.25

# Vector DB (Qdrant - self-hosted)
qdrant-client==1.16.2

# Redis cache
redis==5.0.1

# PostgreSQL async driver
asyncpg==0.29.0

# HTTP
httpx==0.25.2
requests==2.31.0

# Google Drive API
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3

# Async
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Production
gunicorn==21.2.0
psycopg2-binary==2.9.11

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
